move-lightnode-from-fbx-to-mtl
